Refactoring User/Admin, rearrange action folder
