+++ action/event/ListEvent.py
@@ -1,4 +1,4 @@
-from .Action import Action
+from ..Action import Action
 from DB_utils import list_available_study_group
 class ListEvent(Action):
     def exec(self, conn, user):
