+++ role/User.py
@@ -1,35 +1,31 @@
 from .Role import Role
+from action.Exit import Exit
+from action.Logout import Logout
+from action.event.CreateEvent import CreateEvent
+from action.event.ListEvent import ListEvent
+from action.event.JoinEvent import JoinEvent
+from action.event.LeaveEvent import LeaveEvent
+from action.ListHistory import ListHistory
+from action.FindCourse import FindCourse
+from action.FindReserved import FindReserved
+from action.ModifyUserInfo import ModifyUserInfo
+
 
 class User(Role):
-    def __init__(self, userid, username, pwd, email, isUser, isAdmin):
-        self.userid = userid
-        self.username = username
-        self.pwd = pwd
-        self.email = email
-        self.isUser = isUser
-        self.isAdmin = isAdmin
+    def __init__(self, userid, username, pwd, email):
+        super().__init__(userid, username, pwd, email)
+
+        self.user_action =  [
+                                CreateEvent("Create Study Event"),
+                                ListEvent("List All Available Study Events"),
+                                JoinEvent("Join Study Event"),
+                                LeaveEvent("Leave Study Event"),
+                                ListHistory("List Study Group History"),
+                                FindCourse("Find Course"),
+                                FindReserved("Find Reserved Classroom"),
+                                ModifyUserInfo("Modify User Information"),
+                                Logout("Logout"),
+                                Exit("Leave System")
+                            ]
         
-        self.role = 'Admin' if isAdmin else 'User'
-        print(f'Create user | userid: {self.userid}, username: {self.username}, pwd: {self.pwd}, email: {self.email}, role: {self.role}')
 
-    def get_info_msg_no_pwd(self):
-        return f'userid: {self.userid}, username: {self.username}, email: {self.email}, role: {self.role}'
-    def get_info_msg(self):
-        return f'userid: {self.userid}, username: {self.username}, pwd: {self.pwd}, email: {self.email}, role: {self.role}'
-    def get_username(self):
-        return self.username
-    def get_userid(self):
-        return self.userid
-    def get_email(self):
-        return self.email
-    def check_isAdmin(self):
-        return self.isAdmin
-    
-
-    def list_action(self):
-        msg = ''
-        for key in self.action_dict:
-            msg = msg + f'[{key}] {self.action_dict[key].get_name()}\n'
-        return msg
-    
-
